[F] Fix another long cast to int issue

+++ src/main/java/org/hydev/veracross/analyzer/database/model/CourseInfoRating.java
@@ -78,13 +78,13 @@ public class CourseInfoRating extends DatabaseModel<CourseInfoRating>
             .setParameter("id_user", user).list());
     }
 
-    public static List<CourseInfoRating> getByCourse(long id_ci)
+    public static List<CourseInfoRating> getByCourse(int id_ci)
     {
         return VADatabase.query(s -> s.createQuery("from CourseInfoRating where id_ci = :id_ci")
             .setParameter("id_ci", id_ci).list());
     }
 
-    public static List<CourseInfoRating> getByUserAndCourse(long user, long id_ci)
+    public static List<CourseInfoRating> getByUserAndCourse(long user, int id_ci)
     {
         return VADatabase.query(s -> s.createQuery("from CourseInfoRating where id_user = :id_user and id_ci = :id_ci")
             .setParameter("id_user", user).setParameter("id_ci", id_ci).list());
